// types/models.ts
export type statusStoreEnum = 'online' | 'pending' | 'suspended';
export type statusAddress = 'draft' | 'archive';
export type statusCartEnum = 'draft' | 'abandoned' | 'completed';
export type badge = 'hot' | 'bestdeals' | 'sale' | 'off' | 'soldout' | 'none';
export type Image = {
  url: string;
};
// ...
export type TypeStoreModel = {
  _id: string;
  user_id: string; //clerk user_id
  name: string;
  description: string;
  logo: string;
  products: TypeProductModel[];
  orderitems: TypeOrderItemModel[];
  slides: TypeSlideModel[];
  subscription: TypeSubscriptionModel;
  status: statusStoreEnum;
  createdAt: Date;
};
// ...
export type TypeProductModel = {
  _id: string;
  featured?: boolean;
  badge?: badge;
  name: string;
  slug: string;
  description: string;
  additionnal: string;
  specification: string;
  store: TypeStoreModel;
  collections: TypeCollectionModel;
  tags: TypeCollectionModel;
  category: TypeCategoryModel;
  subCategories: TypeSubCategoryModel[];
  brand: TypeBrandModel;
  details: TypeDetailModel[];
  questions: TypeQuestionModel[];
  reviews: TypeReviewModel[];
  productVariants: TypeProductVariantModel[];
  images: string[];
  price: number;
  discount: number;
  seoSlug?: string;
  seoDescription?: string;
  seoTitle?: string;
  status: 'draft' | 'publish' | 'archive';
  inventory: 'instock' | 'outstock';
  weight: number;
  sku: string;
  unit: string;
  user_id: string;
  views: number;
  createdAt?: Date;
};

export type TypeProductVariantModel = {
  productId: TypeProductModel;
  name: string;
  color: TypeColorModel;
  colorImages: Image[];
  size?: TypeSizeModel;
  sizeImages?: Image[];
  weight: number;
  inventory: 'instock' | 'outstock';
  sku: string;
  price: number;
  discount: number;
  colorValue?: string;
  status: 'draft' | 'publish' | 'archive';
};

export type TypeReviewModel = {
  _id: string;
  product: TypeProductModel;
  store: TypeStoreModel;
  rating: number;
  review: string;
  likes: string[];
  user: {
    _id: string;
    imageUrl: string;
    fullName: string;
  };
};

export type TypeCategoryModel = {
  _id: string;
  name: string;
  description: string;
  slug: string;
  image: string;
  user_id: string;
  status: 'draft' | 'publish' | 'archive';
  createdAt: Date;
  subCategory: TypeSubCategoryModel[];
};

export type TypeTagModel = {
  _id: string;
  name: string;
  description: string;
  slug: string;
  image: string;
  user_id: string;
  status: 'draft' | 'publish' | 'archive';
  createdAt: Date;
};

export type TypeCollectionModel = {
  _id: string;
  name: string;
  description: string;
  slug: string;
  color: string;
  image: string;
  user_id: string;
  status: 'draft' | 'publish' | 'archive';
  createdAt: Date;
};

export type TypeColorModel = {
  _id?: string;
  name: string;
  description: string;
  slug: string;
  images: Image[];
  user_id: string | null | undefined;
  value?: string;
  store?: TypeStoreModel;
  status: 'draft' | 'publish' | 'archive';
  createdAt?: Date;
};

export type TypeSizeModel = {
  _id?: string;
  name: string;
  description: string;
  slug: string;
  images: Image[];
  user_id: string | null | undefined;
  value?: string;
  store?: TypeStoreModel;
  status: 'draft' | 'publish' | 'archive';
  createdAt?: Date;
};

export type TypeSubCategoryModel = {
  _id: string;
  name: string;
  description: string;
  slug: string;
  image: string;
  user_id: string;
  status: 'draft' | 'publish' | 'archive';
  category: TypeCategoryModel;
  //   subCategory: TypeSubCategoryModel[];
};

export type TypeBrandModel = {
  _id: string;
  name: string;
  description: string;
  slug: string;
  image: string;
  user_id: string;
  status: 'draft' | 'publish' | 'archive';
  createdAt: Date;
};
// ...
export type TypeSlideModel = {
  _id: string;
  name: string;
  description: string;
  slug: string;
  image: string;
  user_id: string;
  status: 'draft' | 'publish' | 'archive';
  slideItem: string[];
  createdAt: Date;
};

export type TypePageModel = {
  _id: string;
  icon: string;
  name: string;
  description: string;
  slug: string;
  image: string;
  user_id: string;
  status: 'draft' | 'publish' | 'archive';
  createdAt: Date;
};

export type TypeDetailModel = {
  _id: string;
};
export type TypeSubProductModel = {
  _id: string;
  sku: string;
  style: TypeStyleModel;
  options: TypeOptionsModel[];
  subProduct: TypeSubProductModel;
};

export type TypeQuestionModel = { _id: string };
// ...
export type TypeOrderItemModel = {
  _id: string;
  status:
    | 'pending'
    | 'processing'
    | 'onhold'
    | 'completed'
    | 'refunded'
    | 'failed';
  cartItem: TypeCartItemModel;
  order: TypeOrderModel;
  shipping: TypeShippingModel;
  address: TypeAddressModel;
  store: TypeStoreModel;
  shippingAmount: number;
  delivered: boolean;
  earning: number;
  paidAmount: number;
  trackorder: TypeTrackOrderModel;
  createdAt: Date;
};
export type TypeOptionsModel = { _id: string };
export type TypeStyleModel = {
  _id: string;
  color: string;
  image: string;
};

export type TypeSubscriptionModel = {
  store: TypeStoreModel;
  payments: string[];
  startDate: Date;
  endDate?: Date | undefined;
  status: 'active' | 'ended' | 'cancelled';
  type: 'free' | 'pro';
  user_id: string;
};